% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRchi.R
\name{LRchi}
\alias{LRchi}
\title{Compute Likelihood Ratio Chi-square for Binomial Logistic Regression with up to 10 predictors}
\usage{
LRchi(
  data = NULL,
  y = NULL,
  x1 = NULL,
  x2 = NULL,
  x3 = NULL,
  x4 = NULL,
  x5 = NULL,
  x6 = NULL,
  x7 = NULL,
  x8 = NULL,
  x9 = NULL,
  x10 = NULL,
  numpred = NULL
)
}
\arguments{
\item{data}{name of your datafile, loaded}

\item{y}{dependent variable name}

\item{x1}{first predictor variable name}

\item{x2}{second predictor variable name}

\item{x3}{third predictor variable name}

\item{x4}{fourth predictor variable name}

\item{x5}{fifth predictor variable name}

\item{x6}{sixth predictor variable name}

\item{x7}{seventh predictor variable name}

\item{x8}{eighth predictor variable name}

\item{x9}{ninth predictor variable name}

\item{x10}{tenth predictor variable name}

\item{numpred}{number of predictors}
}
\value{
Likelihood Ratio Chi-square for Logistic Regression with up to 10 predictors
}
\description{
Compute Likelihood Ratio Chi-square for Binomial Logistic Regression with up to 10 predictors
}
\examples{
LRchi(data=testlog, y=y, x1=x1, x2=x2,numpred=2)
}
