% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mahal.R
\name{Mahal}
\alias{Mahal}
\title{Compute Mahalanobis Distance for Multiple Regression}
\usage{
Mahal(model = NULL, pred = NULL, values = 5)
}
\arguments{
\item{model}{name of model}

\item{pred}{number of predictors}

\item{values}{number of Mahal values to print (highest values). Default is 10}
}
\value{
Mahalanobis Distance to detect MV outliers
}
\description{
Compute Mahalanobis Distance for Multiple Regression
}
\examples{
\donttest{mymodel<-lm(y~x1+x2+x3+x4, testreg)}
\donttest{Mahal(model=mymodel, pred=5, values = 10)}

}
