% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo.R
\name{pseudo}
\alias{pseudo}
\title{Pseudo R-square Values for Binomial Logistic Regression}
\usage{
pseudo(model = NULL)
}
\arguments{
\item{model}{name of model}
}
\value{
Pseudo R-square Values for Logistic Regression
}
\description{
Pseudo R-square Values for Binomial Logistic Regression
}
\examples{
\donttest{mymodel<-glm(dv~iv1+iv2+iv3+iv4, testlog,family = binomial())}
\donttest{pseudo(model=mymodel)}

}
