% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psuedo.R
\name{psuedo}
\alias{psuedo}
\title{Psuedo R-square Values for Binomial Logistic Regression}
\usage{
psuedo(model = NULL)
}
\arguments{
\item{model}{name of model}
}
\value{
Psuedo R-square Values for Logistic Regression
}
\description{
Psuedo R-square Values for Binomial Logistic Regression
}
\examples{
\donttest{mymodel<-glm(dv~iv1+iv2+iv3+iv4, testlog,family = binomial())}
\donttest{psuedo(model=mymodel)}

}
