% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2change.R
\name{R2change}
\alias{R2change}
\title{R-square change for Hierarchical Multiple Regression}
\usage{
R2change(model1 = NULL, model2 = NULL)
}
\arguments{
\item{model1}{first regression model}

\item{model2}{second regression model}
}
\description{
R-square change for Hierarchical Multiple Regression
}
