% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample1}
\alias{sample1}
\title{sample1}
\format{
A data frame with 713 rows and 14 variables:
\describe{
  \item{id}{ID}
  \item{AA_DV}{Dependent variable}
  \item{AA_Initial}{1st predictor}
  \item{merit}{2nd predictor}
  \item{discrim}{3rd predictor}
  \item{pers_exp}{4th predictor}
  \item{liberal}{5th predictor}
  \item{female}{6th predictor}
  \item{economic}{7th predictor}
  \item{div_part}{8th predictor}
  \item{pers_expN}{9th predictor}
  \item{liberalN}{10th predictor}
  \item{femaleN}{11th predictor}
  \item{economicN}{12th predictor}
  }
}
\usage{
sample1
}
\description{
A dataset to test comparisons between independent coefficients
}
\keyword{datasets}
