% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample2}
\alias{sample2}
\title{sample2}
\format{
A data frame with 164 rows and 11 variables:
\describe{
\item{AA_DV}{Dependent variable}
\item{div_part}{1st predictor}
\item{merit}{2nd predictor}
\item{discrim}{3rd predictor}
}
}
\usage{
sample2
}
\description{
A dataset to test logistic regression functions
}
\keyword{datasets}
