% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testlog}
\alias{testlog}
\title{testlog}
\format{
A data frame with 164 rows and 11 variables:
\describe{
  \item{dv}{DV}
  \item{iv1}{1st predictor}
  \item{iv2}{2nd predictor}
  \item{iv3}{3rd predictor}
  \item{iv4}{4th predictor}
  \item{iv5}{5th predictor}
  \item{iv6}{6th predictor}
  \item{iv7}{7th predictor}
  \item{iv8}{8th predictor}
  \item{iv9}{9th predictor}
  \item{iv10}{10th predictor}
  }
}
\usage{
testlog
}
\description{
A dataset to test logistic regression functions
}
\keyword{datasets}
