% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testreg}
\alias{testreg}
\title{testreg}
\format{
A data frame with 1000 rows and 6 variables:
\describe{
  \item{y}{DV}
  \item{x1}{1st predictor}
  \item{x2}{2nd predictor}
  \item{x3}{3rd predictor}
  \item{x4}{4th predictor}
  \item{x5}{5th predictor}
  }
}
\usage{
testreg
}
\description{
A dataset to test regression functions
}
\keyword{datasets}
