% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolerance.R
\name{tolerance}
\alias{tolerance}
\title{Compute tolerance for Multiple Regression}
\usage{
tolerance(model = NULL)
}
\arguments{
\item{model}{name of model}
}
\value{
Tolerance for MR
}
\description{
Compute tolerance for Multiple Regression
}
\examples{
mymodel<-lm(y~x1+x2+x3+x4+x5, data=testreg)
tolerance(model=mymodel)
}
