% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parts.R
\name{parts}
\alias{parts}
\title{Compute squared partials and semi partial correlatoins for Multiple Regression}
\usage{
parts(model = NULL, pred = NULL)
}
\arguments{
\item{model}{name of model}

\item{pred}{number of predictors}
}
\value{
Squared partial and semipartial correlations for MRC with up to 10 predictors
}
\description{
Compute squared partials and semi partial correlatoins for Multiple Regression
}
\examples{
#'mymodel<-lm(y~x1+x2+x3+x4+x5+x6+x7+x8+x9+x10, data=testreg)
parts(model=mymodel, pred=10)

}
