% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parts.R
\name{parts}
\alias{parts}
\title{Compute squared semi partial correlatoins for Multiple Regression}
\usage{
parts(model = NULL, pred = NULL)
}
\arguments{
\item{model}{name of model}

\item{pred}{number of predictors}
}
\value{
Squared semipartial correlations for MRC with up to 10 predictors
}
\description{
Compute squared semi partial correlatoins for Multiple Regression
}
\examples{
\donttest{mymodel<-lm(y~x1+x2+x3+x4+x5, data=testreg)}
\donttest{parts(model=mymodel, pred=5)}

}
