% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psuedo.R
\name{psuedo}
\alias{psuedo}
\title{Psuedo R-square Values for Binomial Logistic Regression}
\usage{
psuedo(model = NULL)
}
\arguments{
\item{model}{name of model}
}
\value{
Psuedo R-square Values for Logistic Regression
}
\description{
Psuedo R-square Values for Binomial Logistic Regression
}
\examples{
mymodel<-glm(y~x1+x2+x3+x4, testlog,family = binomial())
psuedo(model=mymodel)

}
